add new migrations system (#10312)

+++ packages/db/src/core/cli/migration-queries.ts
@@ -2,11 +2,11 @@ import deepDiff from 'deep-diff';
 import { SQLiteAsyncDialect } from 'drizzle-orm/sqlite-core';
 import * as color from 'kleur/colors';
 import { customAlphabet } from 'nanoid';
-import prompts from 'prompts';
 import { hasPrimaryKey } from '../../runtime/index.js';
 import {
 	getCreateIndexQueries,
 	getCreateTableQuery,
+	getDropTableIfExistsQuery,
 	getModifiers,
 	getReferencesConfig,
 	hasDefault,
@@ -18,6 +18,7 @@ import {
 	type ColumnType,
 	type DBColumn,
 	type DBColumns,
+	type DBConfig,
 	type DBSnapshot,
 	type DBTable,
 	type DBTables,
@@ -28,49 +29,39 @@ import {
 	type TextColumn,
 	columnSchema,
 } from '../types.js';
+import { getRemoteDatabaseUrl } from '../utils.js';
+import { RENAME_COLUMN_ERROR, RENAME_TABLE_ERROR } from '../errors.js';
 
 const sqlite = new SQLiteAsyncDialect();
 const genTempTableName = customAlphabet('abcdefghijklmnopqrstuvwxyz', 10);
 
-/** Dependency injected for unit testing */
-type AmbiguityResponses = {
-	collectionRenames: Record<string, string>;
-	columnRenames: {
-		[collectionName: string]: Record<string, string>;
-	};
-};
-
 export async function getMigrationQueries({
 	oldSnapshot,
 	newSnapshot,
-	ambiguityResponses,
 }: {
 	oldSnapshot: DBSnapshot;
 	newSnapshot: DBSnapshot;
-	ambiguityResponses?: AmbiguityResponses;
 }): Promise<{ queries: string[]; confirmations: string[] }> {
 	const queries: string[] = [];
 	const confirmations: string[] = [];
-	let added = getAddedCollections(oldSnapshot, newSnapshot);
-	let dropped = getDroppedCollections(oldSnapshot, newSnapshot);
-	if (!isEmpty(added) && !isEmpty(dropped)) {
-		const resolved = await resolveCollectionRenames(added, dropped, ambiguityResponses);
-		added = resolved.added;
-		dropped = resolved.dropped;
-		for (const { from, to } of resolved.renamed) {
-			const renameQuery = `ALTER TABLE ${sqlite.escapeName(from)} RENAME TO ${sqlite.escapeName(
-				to
-			)}`;
-			queries.push(renameQuery);
-		}
+	const addedCollections = getAddedCollections(oldSnapshot, newSnapshot);
+	const droppedTables = getDroppedCollections(oldSnapshot, newSnapshot);
+	const notDeprecatedDroppedTables = Object.fromEntries(
+		Object.entries(droppedTables).filter(([, table]) => !table.deprecated)
+	);
+	if (!isEmpty(addedCollections) && !isEmpty(notDeprecatedDroppedTables)) {
+		throw new Error(
+			RENAME_TABLE_ERROR(Object.keys(addedCollections)[0], Object.keys(notDeprecatedDroppedTables)[0])
+		);
 	}
 
-	for (const [collectionName, collection] of Object.entries(added)) {
+	for (const [collectionName, collection] of Object.entries(addedCollections)) {
+		queries.push(getDropTableIfExistsQuery(collectionName));
 		queries.push(getCreateTableQuery(collectionName, collection));
 		queries.push(...getCreateIndexQueries(collectionName, collection));
 	}
 
-	for (const [collectionName] of Object.entries(dropped)) {
+	for (const [collectionName] of Object.entries(droppedTables)) {
 		const dropQuery = `DROP TABLE ${sqlite.escapeName(collectionName)}`;
 		queries.push(dropQuery);
 	}
@@ -78,6 +69,19 @@ export async function getMigrationQueries({
 	for (const [collectionName, newCollection] of Object.entries(newSnapshot.schema)) {
 		const oldCollection = oldSnapshot.schema[collectionName];
 		if (!oldCollection) continue;
+		const addedColumns = getAdded(oldCollection.columns, newCollection.columns);
+		const droppedColumns = getDropped(oldCollection.columns, newCollection.columns);
+		const notDeprecatedDroppedColumns = Object.fromEntries(
+			Object.entries(droppedColumns).filter(([key, col]) => !col.schema.deprecated)
+		);
+		if (!isEmpty(addedColumns) && !isEmpty(notDeprecatedDroppedColumns)) {
+			throw new Error(
+				RENAME_COLUMN_ERROR(
+					`${collectionName}.${Object.keys(addedColumns)[0]}`,
+					`${collectionName}.${Object.keys(notDeprecatedDroppedColumns)[0]}`
+				)
+			);
+		}
 		const result = await getCollectionChangeQueries({
 			collectionName,
 			oldCollection,
@@ -93,18 +97,16 @@ export async function getCollectionChangeQueries({
 	collectionName,
 	oldCollection,
 	newCollection,
-	ambiguityResponses,
 }: {
 	collectionName: string;
 	oldCollection: DBTable;
 	newCollection: DBTable;
-	ambiguityResponses?: AmbiguityResponses;
 }): Promise<{ queries: string[]; confirmations: string[] }> {
 	const queries: string[] = [];
 	const confirmations: string[] = [];
 	const updated = getUpdatedColumns(oldCollection.columns, newCollection.columns);
-	let added = getAdded(oldCollection.columns, newCollection.columns);
-	let dropped = getDropped(oldCollection.columns, newCollection.columns);
+	const added = getAdded(oldCollection.columns, newCollection.columns);
+	const dropped = getDropped(oldCollection.columns, newCollection.columns);
 	/** Any foreign key changes require a full table recreate */
 	const hasForeignKeyChanges = Boolean(
 		deepDiff(oldCollection.foreignKeys, newCollection.foreignKeys)
@@ -120,12 +122,7 @@ export async function getCollectionChangeQueries({
 			confirmations,
 		};
 	}
-	if (!hasForeignKeyChanges && !isEmpty(added) && !isEmpty(dropped)) {
-		const resolved = await resolveColumnRenames(collectionName, added, dropped, ambiguityResponses);
-		added = resolved.added;
-		dropped = resolved.dropped;
-		queries.push(...getColumnRenameQueries(collectionName, resolved.renamed));
-	}
+
 	if (
 		!hasForeignKeyChanges &&
 		isEmpty(updated) &&
@@ -207,116 +204,6 @@ function getChangeIndexQueries({
 	return queries;
 }
 
-type Renamed = Array<{ from: string; to: string }>;
-
-async function resolveColumnRenames(
-	collectionName: string,
-	mightAdd: DBColumns,
-	mightDrop: DBColumns,
-	ambiguityResponses?: AmbiguityResponses
-): Promise<{ added: DBColumns; dropped: DBColumns; renamed: Renamed }> {
-	const added: DBColumns = {};
-	const dropped: DBColumns = {};
-	const renamed: Renamed = [];
-
-	for (const [columnName, column] of Object.entries(mightAdd)) {
-		let oldColumnName = ambiguityResponses
-			? ambiguityResponses.columnRenames[collectionName]?.[columnName] ?? '__NEW__'
-			: undefined;
-		if (!oldColumnName) {
-			const res = await prompts(
-				{
-					type: 'select',
-					name: 'columnName',
-					message:
-						'New column ' +
-						color.blue(color.bold(`${collectionName}.${columnName}`)) +
-						' detected. Was this renamed from an existing column?',
-					choices: [
-						{ title: 'New column (not renamed from existing)', value: '__NEW__' },
-						...Object.keys(mightDrop)
-							.filter((key) => !(key in renamed))
-							.map((key) => ({ title: key, value: key })),
-					],
-				},
-				{
-					onCancel: () => {
-						process.exit(1);
-					},
-				}
-			);
-			oldColumnName = res.columnName as string;
-		}
-
-		if (oldColumnName === '__NEW__') {
-			added[columnName] = column;
-		} else {
-			renamed.push({ from: oldColumnName, to: columnName });
-		}
-	}
-	for (const [droppedColumnName, droppedColumn] of Object.entries(mightDrop)) {
-		if (!renamed.find((r) => r.from === droppedColumnName)) {
-			dropped[droppedColumnName] = droppedColumn;
-		}
-	}
-
-	return { added, dropped, renamed };
-}
-
-async function resolveCollectionRenames(
-	mightAdd: DBTables,
-	mightDrop: DBTables,
-	ambiguityResponses?: AmbiguityResponses
-): Promise<{ added: DBTables; dropped: DBTables; renamed: Renamed }> {
-	const added: DBTables = {};
-	const dropped: DBTables = {};
-	const renamed: Renamed = [];
-
-	for (const [collectionName, collection] of Object.entries(mightAdd)) {
-		let oldCollectionName = ambiguityResponses
-			? ambiguityResponses.collectionRenames[collectionName] ?? '__NEW__'
-			: undefined;
-		if (!oldCollectionName) {
-			const res = await prompts(
-				{
-					type: 'select',
-					name: 'collectionName',
-					message:
-						'New collection ' +
-						color.blue(color.bold(collectionName)) +
-						' detected. Was this renamed from an existing collection?',
-					choices: [
-						{ title: 'New collection (not renamed from existing)', value: '__NEW__' },
-						...Object.keys(mightDrop)
-							.filter((key) => !(key in renamed))
-							.map((key) => ({ title: key, value: key })),
-					],
-				},
-				{
-					onCancel: () => {
-						process.exit(1);
-					},
-				}
-			);
-			oldCollectionName = res.collectionName as string;
-		}
-
-		if (oldCollectionName === '__NEW__') {
-			added[collectionName] = collection;
-		} else {
-			renamed.push({ from: oldCollectionName, to: collectionName });
-		}
-	}
-
-	for (const [droppedCollectionName, droppedCollection] of Object.entries(mightDrop)) {
-		if (!renamed.find((r) => r.from === droppedCollectionName)) {
-			dropped[droppedCollectionName] = droppedCollection;
-		}
-	}
-
-	return { added, dropped, renamed };
-}
-
 function getAddedCollections(oldCollections: DBSnapshot, newCollections: DBSnapshot): DBTables {
 	const added: DBTables = {};
 	for (const [key, newCollection] of Object.entries(newCollections.schema)) {
@@ -331,20 +218,6 @@ function getDroppedCollections(oldCollections: DBSnapshot, newCollections: DBSna
 		if (!(key in newCollections.schema)) dropped[key] = oldCollection;
 	}
 	return dropped;
-}
-
-function getColumnRenameQueries(unescapedCollectionName: string, renamed: Renamed): string[] {
-	const queries: string[] = [];
-	const collectionName = sqlite.escapeName(unescapedCollectionName);
-
-	for (const { from, to } of renamed) {
-		const q = `ALTER TABLE ${collectionName} RENAME COLUMN ${sqlite.escapeName(
-			from
-		)} TO ${sqlite.escapeName(to)}`;
-		queries.push(q);
-	}
-
-	return queries;
 }
 
 /**
@@ -551,4 +424,30 @@ type DBColumnWithDefault =
 
 function hasRuntimeDefault(column: DBColumn): column is DBColumnWithDefault {
 	return !!(column.schema.default && isSerializedSQL(column.schema.default));
+}
+
+export async function getProductionCurrentSnapshot({
+	appToken,
+}: {
+	appToken: string;
+}): Promise<DBSnapshot> {
+	const url = new URL('/db/schema', getRemoteDatabaseUrl());
+
+	const response = await fetch(url, {
+		method: 'POST',
+		headers: new Headers({
+			Authorization: `Bearer ${appToken}`,
+		}),
+	});
+	const result = await response.json();
+	return result.data;
+}
+
+export function createCurrentSnapshot({ tables = {} }: DBConfig): DBSnapshot {
+	const schema = JSON.parse(JSON.stringify(tables));
+	return { experimentalVersion: 1, schema };
+}
+
+export function createEmptySnapshot(): DBSnapshot {
+	return { experimentalVersion: 1, schema: {} };
 }
